Split rules from smart playlist into new file
